@@ -1,6 +1,6 @@
 module github.com/kubeshop/tracetest/server
 
-go 1.18
+go 1.20
 
 replace k8s.io/client-go => k8s.io/client-go v0.18.0
 
