@@ -37,7 +37,7 @@ type (
 	HttpClientConfig struct {
 		Url        string
 		Headers    map[string]string
-		TLSSetting configtls.TLSClientSetting
+		TLSSetting configtls.TLSClientSetting `mapstructure:"tls"`
 	}
 
 	OTELCollectorConfig struct {
@@ -103,6 +103,10 @@ var otlpBasedDataStores = []DataStoreType{
 	DataStoreTypeNewRelic,
 	DataStoreTypeLighStep,
 	DataStoreTypeDataDog,
+}
+
+func (ds DataStore) HasID() bool {
+	return ds.ID != ""
 }
 
 func (ds DataStore) Validate() error {
